Move releaseTag out of benchmark-store
The release tag of the running build is no longer benchmark
property: it moves to src/lib/server/release-tag.ts so /api/v1
version can serve the same definition without reaching into the
benchmark module. No behavior change — same override, same
fallback, same tests, now pointed at the new home.
