+++ src/lib/server/benchmark-store.ts
@@ -14,17 +14,13 @@
  * deterministic — see the header of `content-quality.ts`.
  */
 import type { SupabaseClient } from '@supabase/supabase-js';
-import { env } from '$env/dynamic/private';
 import {
   CONTENT_SCORER_VERSION,
   checkValues,
   scoreContentQuality
 } from '$lib/server/content-quality';
 import { summarize, type Sample } from '$lib/server/benchmark';
-import { formatRelease } from '$lib/release';
-// Only `version` is read. JSON imports are enabled (tsconfig `resolveJsonModule`) and the named
-// import lets the bundler drop the rest of the manifest.
-import { version as pkgVersion } from '../../../package.json';
+import { releaseTag } from '$lib/server/release-tag';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 type AnyRec = Record<string, any>;
@@ -39,23 +35,10 @@ export const BENCHMARK_BATCH = 200;
 export const REPETITION_WINDOW = 15;
 
 /**
- * The build that generated the content, as `<semver>+<commit>` — the same tag `svelte.config.js`
- * hands to `$app/environment`, composed here from the same two inputs.
- *
- * It is NOT read from `$app/environment`: no module under `src/lib/server/` imports it, because
- * that would make this file untestable. Both sides derive the tag from `formatRelease`, so the two
- * cannot drift in shape — only in timing (this one reads the commit at runtime, Kit bakes it at
- * build; on Vercel they are the same value).
- *
- * `APP_RELEASE` overrides everything for self-hosted or scripted runs. Without a tag a sample
- * cannot be attributed to a change, so the fallback is a loud `dev`, never null.
+ * The build that generated the content, as `<semver>+<commit>`.
+ * The definition lives in `src/lib/server/release-tag.ts`, shared with
+ * `/api/v1/version` — one place, so the two cannot drift.
  */
-export function releaseTag(): string {
-  const explicit = String(env.APP_RELEASE ?? '').trim();
-  if (explicit) return explicit.slice(0, 60);
-  const commit = String(env.VERCEL_GIT_COMMIT_SHA ?? '').trim().slice(0, 12);
-  return formatRelease(pkgVersion, commit || 'dev');
-}
 
 export type PostRow = {
   id: string;
